Keep both 388 schema notes and main CI/lint docs. Needed so PR 3 builds on Money + currency schema.

+++ crates/db/src/recurring_transactions/adopt.rs
@@ -99,7 +99,7 @@ fn transaction_matches_review(
 
     transaction.transaction_date == input.expected_transaction_date
         && description == Some(input.template.description.as_str())
-        && transaction.amount == input.template.amount
+        && transaction.amount == i64::from(input.template.amount)
         && transaction.transaction_type == input.template.transaction_type
         && transaction.transaction_category_id == input.template.transaction_category_id
         && notes == input.template.notes.as_deref()
